add stubs of service and checkout service methods implementation

+++ src/main/groovy/org/jsalazar/checkoutservice/client/interfaces/FinancialService.groovy
@@ -3,7 +3,7 @@ package org.jsalazar.checkoutservice.client.interfaces
 import org.jsalazar.checkoutservice.common.dto.Card
 
 interface FinancialService {
-    Long charge(Card card, int amount)
+    Long charge(Card card, double amount)
 
     void voidTransaction(Long transactionId)
